Added category ownership

+++ MyStoreBack/Business logic/Category/CategoryService.cs
@@ -1,8 +1,12 @@
+using System.Security.Claims;
 using AutoMapper;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 using MyStoreBack.Business_logic.Files;
+using MyStoreBack.Constants;
 using MyStoreBack.Data.Context;
 using MyStoreBack.Data.Entity;
+using MyStoreBack.Data.Entity.Identity;
 using MyStoreBack.Models.Category;
 
 namespace MyStoreBack.Business_logic.Category;
@@ -10,50 +14,78 @@ namespace MyStoreBack.Business_logic.Category;
 public class CategoryService : ICategoryService
 {
     private readonly IPictureService _pictureService;
-    private StoreDbContext Context { get; }
-    private IMapper Mapper { get; }
+    private readonly StoreDbContext _context;
+    private readonly IMapper _mapper;
+    private readonly UserManager<UserEntity> _userManager;
 
     public CategoryService(
         StoreDbContext context,
         IMapper mapper,
-        IPictureService pictureService
+        IPictureService pictureService,
+        UserManager<UserEntity> userManager
     )
     {
         _pictureService = pictureService;
-        Context = context;
-        Mapper = mapper;
+        _context = context;
+        _mapper = mapper;
+        _userManager = userManager;
     }
+
+    private async Task<UserEntity?> GetCurrentUser(ClaimsPrincipal user)
+        => await _userManager.GetUserAsync(user);
     
-    public async Task<IEnumerable<CategoryModel>> GetAll()
+    private bool IsAdmin(UserEntity? user)
+        => user != null && _userManager.IsInRoleAsync(user, Roles.Admin).Result;
+    
+    public async Task<IEnumerable<CategoryModel>> GetAll(ClaimsPrincipal user)
     {
-        var categories = await Context.Categories.ToListAsync();
-        var mapped = Mapper.Map<IEnumerable<CategoryModel>>(categories);
-
-        return mapped;
-    }
-
-    public async Task<CategoryModel> GetById(int id)
-    {
-        var cat = await Context.Categories.FindAsync(id);
-        var mapped = Mapper.Map<CategoryModel>(cat);
+        var userEntity = await GetCurrentUser(user);
+        if (userEntity is null) 
+            throw new UnauthorizedAccessException("User not found.");
         
+        var categories = await _context.Categories
+            .Where(x => x.UserId == userEntity.Id || IsAdmin(userEntity)).ToListAsync();
+        var mapped = _mapper.Map<IEnumerable<CategoryModel>>(categories);
+
         return mapped;
     }
 
-    public async Task<CategoryModel?> Create(CategoryCreateModel model)
+    public async Task<CategoryModel> GetById(int id, ClaimsPrincipal user)
+    {
+        var userEntity = await GetCurrentUser(user);
+        if (userEntity is null) 
+            throw new UnauthorizedAccessException("User not found.");
+        
+        var cat = await _context.Categories.FindAsync(id);
+        
+        if (cat is null || (userEntity?.Id != cat?.UserId && !IsAdmin(userEntity))) 
+            throw new Exception("Category not found.");
+        
+        var mapped = _mapper.Map<CategoryModel>(cat);
+        return mapped;
+    }
+
+    public async Task<CategoryModel?> Create(CategoryCreateModel model, ClaimsPrincipal user)
     {
         try
         {
-            var cat = Mapper.Map<CategoryEntity>(model);
+            var userEntity = await GetCurrentUser(user);
+            if (userEntity is null) 
+                throw new UnauthorizedAccessException();
+            
+            var cat = _mapper.Map<CategoryEntity>(model);
             
             // Save image
             string imageUrl = await _pictureService.Save(model.Image);
             cat.ImageUrl = imageUrl;
             
-            await Context.AddAsync(cat);
-            await Context.SaveChangesAsync();
+            // Add UserId
+            cat.UserId = userEntity.Id;
             
-            var mapped = Mapper.Map<CategoryModel>(cat);
+            await _context.AddAsync(cat);
+            await _context.SaveChangesAsync();
+            
+            var mapped = _mapper.Map<CategoryModel>(cat);
             return mapped;
         }
         catch
@@ -62,13 +94,21 @@ public class CategoryService : ICategoryService
         }
     }
 
-    public async Task<CategoryModel?> Update(CategoryUpdateModel model)
+    public async Task<CategoryModel?> Update(CategoryUpdateModel model, ClaimsPrincipal user)
     {
         try
         {
-            var cat = await Context.Categories.FindAsync(model.Id) 
+            var userEntity = await GetCurrentUser(user);
+            if (userEntity is null) 
+                throw new UnauthorizedAccessException();
+            
+            var cat = await _context.Categories.FindAsync(model.Id) 
                       ?? throw new Exception("Category not found.");
-            Mapper.Map(model, cat);
+            
+            if (cat.UserId != userEntity.Id && !IsAdmin(userEntity))
+                throw new UnauthorizedAccessException();
+            
+            _mapper.Map(model, cat);
             
             // Unlinking files and saving a new picture
             if (model.Image is not null) {
@@ -77,10 +117,10 @@ public class CategoryService : ICategoryService
                 cat.ImageUrl = newImageUrl;
             }
 
-            Context.Update(cat);
-            await Context.SaveChangesAsync();
+            _context.Update(cat);
+            await _context.SaveChangesAsync();
 
-            var mapped = Mapper.Map<CategoryModel>(cat);
+            var mapped = _mapper.Map<CategoryModel>(cat);
             return mapped;
         }
         catch
@@ -89,16 +129,21 @@ public class CategoryService : ICategoryService
         }
     }
 
-    public async Task<bool> Delete(int id)
+    public async Task<bool> Delete(int id, ClaimsPrincipal user)
     {
-        var cat = await Context.Categories.FindAsync(id);
+        var cat = await _context.Categories.FindAsync(id);
         if (cat is null) return false;
+        
+        var userEntity = await GetCurrentUser(user);
+        if (userEntity is null ||
+            (cat.UserId != userEntity.Id && !IsAdmin(userEntity)))
+            return false;
         
         // Unlinking files
         _pictureService.RemoveByUrl(cat.ImageUrl);
             
-        Context.Remove(cat);
-        await Context.SaveChangesAsync();
+        _context.Remove(cat);
+        await _context.SaveChangesAsync();
             
         return true;
     }
